+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## HEAD
+* Added _README.md file to the template directory. Needs to be written for the user.
+* A new application logic to copy over the necessary files for the app.
+* Minor restructuring of the index.js file.
+* Rewrote the descriptions that needed an update. ([#11](https://github.com/sondr3/generator-jekyllized/issues/11))
 * Fix the pathing issues with cssExDir, jsExDir etc. ([#10](https://github.com/sondr3/generator-jekyllized/issues/10))
 * Added support for installing Google Analytics during the installation. ([#9](https://github.com/sondr3/generator-jekyllized/issues/9))
 * Moved assets into their own folder by default. ([#8](https://github.com/sondr3/generator-jekyllized/issues/8)) & ([#3](https://github.com/sondr3/generator-jekyllized/issues/3))
@@ -7,7 +11,7 @@
 * Renamed projectname into appname so it generates correctly. ([#2](https://github.com/sondr3/generator-jekyllized/issues/2))
 
 ## 0.0.2 (January 24, 2014)
-* Restructured the index.js file so it asks for information about the project first and then the owner
+* Restructured the index.js file so it asks for information about the project first and then the owner. ([#1](https://github.com/sondr3/generator-jekyllized/issues/1))
 * Added fields for a short biography, Twitter, Google Plus and tagline for the project and owner
 * Renamed some of the fields for easier understanding